+++ packages/compat/config/chain.ts
@@ -1,28 +1,8 @@
-import { ParachainId } from '@zenlink-interface/chain'
-
-export const AMM_ENABLED_NETWORKS = [
-  ParachainId.ASTAR,
-  ParachainId.MOONRIVER,
-  ParachainId.MOONBEAM,
-]
-
-export const EVM_NETWORKS = [
-  ParachainId.ASTAR,
-  ParachainId.MOONRIVER,
-  ParachainId.MOONBEAM,
-]
-
-export const SUBSTRATE_NETWORKS = [
-  ParachainId.BIFROST_KUSAMA,
-  ParachainId.BIFROST_POLKADOT,
-]
-
-export const SUPPORTED_CHAIN_IDS = Array.from(new Set([...AMM_ENABLED_NETWORKS]))
-
-export function isEvmNetwork(chainId: ParachainId) {
-  return EVM_NETWORKS.includes(chainId)
-}
-
-export function isSubstrateNetwork(chainId: ParachainId) {
-  return SUBSTRATE_NETWORKS.includes(chainId)
-}
+export {
+  AMM_ENABLED_NETWORKS,
+  EVM_NETWORKS,
+  isEvmNetwork,
+  isSubstrateNetwork,
+  SUBSTRATE_NETWORKS,
+  SUPPORTED_CHAIN_IDS,
+} from '@zenlink-interface/chain'
