+++ packages/graph-client/src/index.ts
@@ -2,3 +2,4 @@ export * from './appolo'
 export * from './queries'
 export * from './resolvers'
 export * from './types'
+export * from './utils'
